Extract "year" into config variable
In fact this is not quite the year, as the URL and all references are 'spa2018' etc rather than '2018'. This might be something I'll change later.

This also extracts the current year out: this is only used in the footer, and updates with the year. I've changed the functionality as the current site design dates from 2016, so the copyright should be a range (http://www.copyrightservice.co.uk/copyright/p03_copyright_notices).

This also changes the page title from "SPA2018 conference" to "spa2018 conference" - not ideal, but I do not think it's worth creating an extra variable for this.