Add lead a session page
This is currently being rewritten by the new chairs and may not even include the navigation sub-headings so I've kept it brief.

This commit also changes the homepage to update the number of years the conference has been running. I've done this manually rather than programmatically because it may be that this text should be rewritten as well.