Use `relative_url` to fix relative URLs
In `da71c2a` I removed the `base` tag and made the links relative so that the site would work locally, but that only works locally because the site is at the root of the domain. On the server, this site sits in a sub-directory (e.g. `spa2018`) so these links, relative to the root, do not work on the server.

This commit fixes that by setting a `baseurl` (currently `/tmp/` but once the site is active that will be the actual directory we want to deploy to), and using Jekyll's `relative_url` filter to make URLs do the right thing.

Documentation here: https://jekyllrb.com/docs/templates/

Useful blog post (though it trails off in the middle of a thought...), here: https://byparker.com/blog/2014/clearing-up-confusion-around-baseurl/.