Pull navigation out of the main content
This commit addresses the issue identified in `04667666` in that the navigation was inside the content block. This pulls the two in-page navigations out into separate includes and uses front-matter variables to identify which nav to include, and the URLs to identify which menu item should be selected (in the previous sites page there is no need for the latter as all the menu items take you off the page).

The `menuactive` class was duplicated on the list item, this was an artefact of the CMS so has been removed.