Make HTML proofer pass
HTML proofer fails because this image does not have an alt tag. In fact, I left it out deliberately because it doesn't do anything for accessbility or for page improvement for there to be an alt-tag here. However, on balance HTML proofer seems to be a good thing, so adding the tag in here.