Add no cookies page
This is linked out to from the login page if people do not have cookies enabled. From there it is `nocookies.html` - I've changed the URL structure to match the others and `CheckSpelling` will find it, so it will work for now, until I've updated the code. This will need some future work but it's outside the current scope.