Deploy to actual site
This commit changes the folder I've been testing on to the actual folder, so once this is merged, all commits to master will be deployed to the live site.

I had to change the variable in three places which is not ideal, but I've not been able to find a way to consolidate that (can you use variables in the travis config?) so this is a story for another day.