Add interactivity page to replace formats page
We want to encourage people to propose sessions which get people involved in
interesting ways, and this page is intended to help them think about that
aspect of their session in a more open-ended way than the old formats page.

Since the formats page won't be used at all this year, remove it entirely - we
expect that approximately no one will have the link to that page since it's
only been live for a short time and we haven't publicised the site during that
time.

Also update the formats page link from CFP to go to the interactivity page.