Update all dates for the 2018 conference
This will give people 7 weeks to submit proposals before the earlier deadline,
and then 4 weeks between that and the final deadline, and just under 4 for
reviewing before the programme meeting. There's a day between the end of the
review period and the programme meeting to give us time to prepare for the
meeting.

I've left it vague as to when we'll announce the programme to give us some
flexibility, but about a month after the programme meeting seems doable.

I've had to guess at the deadline for ticket cancellations - we'll need to
check this with BCS and update it later. For now I've made it the Wednesday
a week and a half before the conference as it was in 2014, when this page was
last updated.