Move dates into new variables
We've got lots of repetition in the content around dates - I think it could do
with some deduplication, but in the meantime getting these dates into
variables means we only need to update them in one place rather than hunting
round the whole site.

The only dates not covered by this change are on the ticket booking page
(which isn't live yet) - the wording there is different and it seems excessive
to add extra variables just for that case rather than rewording it.