Add clarification about deployment
The deployment does not purge on deploy, which means that existing pages that are no longer present in the source are not removed. This is good because we should not be removing pages that have been published as that breaks existing links, and it is necessary because the programme is generated from the scripts part of the site and is deployed in a different way; if this purged on deploy it would delete the programme every time.