Ensure the new conference name isn't too wide
On narrow devices (notably, iPhones 5/5S/SE) the new conference name,
in a 48px face, is just wide enough that it induces horizontal
scrolling. To counter that, let's just reduce the font size back to the
default. This isn't a perfect fix but, given the current layout, it's
probably the best compromise.

It turns out that the location image on the homepage is *also* wide
enough to cause this issue, so I've fixed that.