Standardise the header on all pages
Now we've made the page header smaller (783c4c5), it looks a bit odd to
have the header changing from one page to the next. To rectify that, we
can show the date everywhere, show the header background image
everywhere, and remove the spacing between the conference title and the
date.