Programmatically generate the menubar from config
The current menubar is easy to understand, but isn't easy to make show
the currently active page. It also requires some duplication, making
future markup changes more complicated.

This change moves the control of the navigation structure into the
`_config.yml` (though it could possibly be done another way), and then
uses that structure to generate the menu programmatically. Menu items
are rendered in the order they're stored in the array, and may either
be the ID of a page, or a dictionary (expected to contain both `title`
and `url`, which we output without touching).

The rendering code checks if the menu item is either the current page,
or is mentioned in the current page's `has-nav` attribute and, if true,
marks the menu item as selected.