Add programme includes
This generates a header and footer for the programme. These are so that the preview programme and published programme can have a header and footer like the rest of the site, even though they are generated from within the PHP rather than in Jekyll.

The Jekyll site will still need to be updated to add the programme to the menu.