Remove comments from menu include
This is written almost entirely in liquid templating, and so the comments don't look any different to the code and add to the overall effect of something really complicated that it's hard to understand. Removing the comments mean you can see that there's not actually much code here so you can probably figure out how it works if you need to.