Update the CSS to handle showing the room names
Last year's programme only showed the track headings (which are now the actual room names) on smaller screens; on the desktop it didn't show them. I've updated the programme so that it will show the room names, and this change to the CSS makes sure that only the time is hidden in tracks 2, 3 and 4, not the room name as well.