Remove separate refunds and cancellations page
The relevant details are better shown on the book now page - having a separate
page encourages us to write a lot more than is required, and you want this
information when you are booking.

I've also removed the variable about the last refund date for two reasons - the
only place it appears needs a lot of manual editing each year anyway, and
secondly we might not be issued this date, we just have to find it out when the
tickets are announced.