Correct `if...else` so body class is correct
This was `if... else if...` but in Jekyll it should actually be `if... elsif...`. The error meant that all pages on the site except the homepage were being set as `body class='programme', which had the effect of `<ul>` on the T&Cs page not showing up (in the programme, unordered lists are used structurally).