Create new headers for different programme pages
It turns out that the CSS of the session and user pages needs to be differentiated. In previous years this hasn't been a problem because we didn't include the main programme header in the page (e.g. https://spaconference.org/spa2017/sessions/session715.html).

We've now realised that the session and user pages should have the same header as the programme. Like the programme, these pages are generated from inside the PHP scripts so will need to grab the include from the jekyll site (to ensure it's always up to date with the rest of the site).

However, we can't use the same include as the programme because the user page needs a body class of 'person' to display correctly, and the session page requires no body class (rather than 'programme').

The way to make this easiest to understand is to create extra includes (rather than burying some CSS in the PHP), so this commit creates those extra includes (and the clause to select the right one for the user pages). It also pulls out the repeated code into an include.