Only swap `baseurl` at start of URL
Without the `^` this will swap out the baseurl wherever it appears in the URL. This causes tests to fail locally when they have the site name in the middle, for example links to the programme, which are currently fully qualified. `https://www.spaconference.org/spa2018/programme.html` becomes `https://www.spaconference.org/programme.html`, which doesn't exist.

Some of these we can change by using relative urls, but some we can't affect; for example this year the link to tickets was `https://www.eventbrite.co.uk/e/spa2018-tickets-42752231074` which becomes `https://www.eventbrite.co.uk/e-tickets-42752231074`.

We only care about swapping out the baseurl at the beginning, so this changes the documentation to add `^/` to the regex.