Add deprecation notice
This leaves it in a bit of a mess; with a mix of spaconference-history and spaconference domains, a travis build that doesn't work, etc. However, once it becomes clear what the next steps are for the site and the hosting, we can return to tidy this up, but in the meantime, it is useful to know it's no longer used.